pyproject.toml
[tool.poetry]
name = "simple_todo"
version = "0.1.0"
description = ""
authors = ["Jonas Reetz <jonas@reetz.me>"]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "*"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
